+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## Unreleased
+## 0.6.5 - 2017-08-04
 ### Changed
 - Change Twig templating extension to `.twig`.